성적등록성적목록구현(미완성인 부분들 있음 추후 수정)

+++ src/main/java/com/schoolproject/repository/LectureRepository.java
@@ -1,9 +1,7 @@
 package com.schoolproject.repository;
 
+import com.schoolproject.entity.Lecture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.schoolproject.entity.Lecture;
-
-public interface LectureRepository extends JpaRepository<Lecture, Integer> {
-    // 추가적인 쿼리 메서드를 여기에 정의
+public interface LectureRepository extends JpaRepository<Lecture, Long> {
 }
